first-pass to readd compression support for s3
this commit is a first-pass at readding support for compression to
s3. I found some mention of the code in contrib/deprecated/producer.
However I found the new one did not respect compression at all.

this commit re-adds support for that by modifying the underlying
byteBuffer/batchedFileWriter to now accept a compression flag, which
if set will cause the buffer to gzip compress the stream.